Finalized controller system
Finally, the controller system is done.
EntityScripts now give their options to the entity and send all the
options to the player. The player paints the option dialog and let the
user choose one.
When that happens, the player receive one option and do whatever.
Also done the GUI system and moved all the moving things in the entity
to the Mover script.