Changes to cell system 1
First of all, updated to Unity 5! Yay! :dancers: (solves #10)
Here we go with the first commit of the changes to the cell system.
- The changes for now have been centered in cleaning some of the cell
generating system, creating independence between the systems. Also, the
cell have been semi cleaned (solves #11) till the full clean next update
after the clean.
- Faces are no longer ScriptableObject, reducing the memory cost and no
longer having to think about it's lifecycle.
- CellProperties class now handles all the information neccesary to
reconstruct cell mesh working along with MeshFactory.
- Cell no longer uses its transform to change height.
- CellEditor now displays a handle to change height. (In next updates
it'll be finished, we're working to improve it, but for now it's fully
functional.)
- CellEditor now uses Undo function to backup previous Cell state.
- Games temporary cant be exported this is caused becouse:
* Map now instanciates Cells from Prefab.
* Cells has to indicate unity editor about it's dirty state for changes
to be saved.
- Cell meshes and textures have hashes to be easily searched. This will
enable the way to share meshes and textures.
- Cell properties can be compared to look after the hashes match.
- Cell now have the hability to reconstruct its faces from the mesh
(textures cannot yet extracted and possible it wont be possible). Faces
also have the capacity to extract shared vertex, triangles and uvs from
mesh.

From now on, this is all time I had to work on the project.
I have to announce a recent update fixing issues with events in memory
that will be explained in next commit.