Tons of changes in Mover & movements
- Movers now Have Direction property to switch direction. Also, an
"turn" event with parameter "direction" with type Mover.Direction
(North, East, South & West) can be used. (solves #40)
- Player now perform direction change in case he cannot move to the
selected direction.
- Added a new IsoUnityType to handle Enums in GameEvents.
- Movements have been independized and since they're instanciated they
handle all the property updates.
- There're four types of movements as for now: Normal, Parabolic (as in
a jump), Instant (as in a teleport) and Turn (as in a direction change).
- Having movement independized means that new movement types could be
added to handle specific situations as in a wall crash or in a free
movement situation. (Probably this last one will be implemented in
future).
- Movements can be stopped (the stop will be effective when the current
movement to next cell ends). The stop can be performed through the event
stop. (solves #20). If the movement is caused by event, even if it's
stopped, the EventFinished will be launched.
- In the last commit, ownership have been added, this includes Tag
recognition.